Branch Ruby and MacRuby implementations
The MacRuby stuff is going to start using a lot of things that are MacRuby specific,
such as GCD. Also, the MacRuby version cannot use curb as the HTTP driver.

In the future it may be better to use branches to keep the implementations separate,
and then merge common changes between the them; but I think the changes will be too
different for that to work.

We shall see.