Change JIRAService factory method to have a more natural name
I was debating with myself whether it should imply that the instance is for an
endpoint of if the instance has an endpoint. I still have not decided, but I think
that calling the server an endpoint means that the instance has an endpoint and so
the method name should be instance_with_endpoint.

If anyone wants to debate this, they are more than welcome to log an issue.