Stop using so many factories!
Switch to using constructors for all the JIRA entities. Also update the API to
use the constructors.

The main motivation for this was so that the JIRA entities could more easily be 
abstracted with base classes and still have a sane initialization process; but
it also means that the code base will be smaller and probably faster.