Iterate again on the RemoteAPI code deduplication
Now we have two methods for public RemoteAPI methods to call
that will do all the heavy lifting.

#call is for simple cases where we want a single object,
sometimes a core type and sometimes a JIRA object.

#jira_call is for when we need to package an array of up,
and will always be an array of JIRA objects.

There are still two cases that I would like to abstract away
from the public part of RemoteAPI:
	- JIRA::Entity.new_with_xml call().first
		need to abstract away .new_with_xml and #first
	- SOAPifying string arrays

This iteration reduces the code size in the RemoteAPI module, but at a some cost to performance that I hope to rectify before
jiraSOAP 0.8.